[front] Split code-defined skills by scope
