+++ front/lib/resources/skill/code_defined/global/docx.ts
@@ -1,6 +1,6 @@
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
-import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/shared";
+import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/global/types";
 import { isComputerFeatureEnabled } from "@app/types/shared/feature_flags";
 
 const DOCX_SKILL_INSTRUCTIONS = `# Documents (.docx, .doc)
@@ -128,6 +128,7 @@ Two passes. Both required — never ship after only the structural pass.
 
 export const docxSkill = {
   sId: "docx",
+  kind: "global",
   name: "Documents",
   userFacingDescription:
     "Read, edit, and create text documents (.docx) in the sandbox.",
