+++ front/lib/resources/skill/code_defined/global/mention_users.ts
@@ -3,10 +3,11 @@ import {
   SEARCH_AVAILABLE_USERS_TOOL_NAME,
   USER_MENTIONS_SERVER_NAME,
 } from "@app/lib/api/actions/servers/user_mentions/metadata";
-import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/shared";
+import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/global/types";
 
 export const mentionUsersSkill = {
   sId: "mention_users",
+  kind: "global",
   name: "Mention Users",
   userFacingDescription:
     "Allow agents to mention and notify workspace users in conversations.",
