+++ front/lib/resources/skill/code_defined/global/pptx.ts
@@ -1,6 +1,6 @@
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
-import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/shared";
+import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/global/types";
 import { isComputerFeatureEnabled } from "@app/types/shared/feature_flags";
 
 const PPTX_SKILL_INSTRUCTIONS = `# Slide decks (.pptx)
@@ -674,6 +674,7 @@ three modes.
 
 export const pptxSkill = {
   sId: "pptx",
+  kind: "global",
   name: "Slide decks",
   userFacingDescription: "Read, edit, and create slide presentations",
   agentFacingDescription:
