+++ front/lib/resources/skill/code_defined/global/projects.ts
@@ -6,7 +6,7 @@ import {
   readPodAgentsMdContent,
 } from "@app/lib/api/projects/agents_md";
 import type { Authenticator } from "@app/lib/auth";
-import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/shared";
+import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/global/types";
 import { SpaceResource } from "@app/lib/resources/space_resource";
 import {
   type ConversationWithoutContentType,
@@ -15,6 +15,7 @@ import {
 
 export const projectsSkill = {
   sId: "projects",
+  kind: "global",
   name: "Pods",
   userFacingDescription:
     "Allow agents to create conversations & messages in Pods, leverage Pods knowledge.",
