+++ front/lib/resources/skill/code_defined/global/support.test.ts
