+++ front/lib/resources/skill/code_defined/global/support.ts
@@ -1,5 +1,5 @@
 import { WEB_SEARCH_BROWSE_SERVER_NAME } from "@app/lib/api/actions/servers/web_search_browse/metadata";
-import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/shared";
+import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/global/types";
 
 const SUPPORT_INSTRUCTIONS = `
 Ground every answer on the best available Dust-specific evidence. Start with public Dust surfaces:
@@ -34,6 +34,7 @@ Be explicit about uncertainty. If public sources do not answer the question, say
 
 export const supportSkill = {
   sId: "support",
+  kind: "global",
   name: "Dust Support",
   userFacingDescription:
     "Get help with Dust using public docs, open-source issues, and community knowledge.",
