+++ front/lib/resources/skill/code_defined/global/types.ts
@@ -1,0 +1,1 @@
+export type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/shared";
