+++ front/lib/resources/skill/code_defined/global/workspace_analytics.test.ts
