+++ front/lib/resources/skill/code_defined/global/xlsx.ts
@@ -1,6 +1,6 @@
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
-import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/shared";
+import type { GlobalSkillDefinition } from "@app/lib/resources/skill/code_defined/global/types";
 import { isComputerFeatureEnabled } from "@app/types/shared/feature_flags";
 
 const XLSX_SKILL_INSTRUCTIONS = `# Spreadsheets (.xlsx, .xlsm, .csv, .tsv)
@@ -124,6 +124,7 @@ Also verify zero formula errors. \`#REF!\`, \`#DIV/0!\`, \`#VALUE!\`, \`#N/A\`,
 
 export const xlsxSkill = {
   sId: "xlsx",
+  kind: "global",
   name: "Spreadsheets",
   userFacingDescription: "Read, edit, and create spreadsheets (.xlsx, .csv)",
   agentFacingDescription:
