+++ front/lib/resources/skill/code_defined/system/plan_mode.ts
@@ -7,7 +7,7 @@ import {
 } from "@app/lib/api/actions/servers/plan_mode/metadata";
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
-import type { SystemSkillDefinition } from "@app/lib/resources/skill/code_defined/shared";
+import type { SystemSkillDefinition } from "@app/lib/resources/skill/code_defined/system/types";
 
 const ASK_USER_QUESTION_TOOL_NAME =
   ASK_USER_QUESTION_TOOLS_METADATA.ask_user_question.name;
@@ -46,6 +46,7 @@ Do NOT close to handle revisions; use \`${EDIT_PLAN_TOOL_NAME}\` to iterate inst
 
 export const planModeSkill = {
   sId: "plan_mode",
+  kind: "system",
   name: "Plan Mode",
   userFacingDescription:
     "Let agents maintain a live plan.md the user can follow as work progresses.",
