+++ front/lib/resources/skill/code_defined/system/types.ts
@@ -1,0 +1,1 @@
+export type { SystemSkillDefinition } from "@app/lib/resources/skill/code_defined/shared";
