+++ front/lib/resources/skill/code_defined/system_registry.ts
@@ -1,13 +1,13 @@
 import type { Authenticator } from "@app/lib/auth";
-import { discoverKnowledgeSkill } from "@app/lib/resources/skill/code_defined/discover_knowledge";
-import { discoverSkillsSkill } from "@app/lib/resources/skill/code_defined/discover_skills";
-import { discoverToolsSkill } from "@app/lib/resources/skill/code_defined/discover_tools";
-import { planModeSkill } from "@app/lib/resources/skill/code_defined/plan_mode";
 import {
   ensureUniqueSIds,
   filterSkillDefinitions,
   type SystemSkillDefinition,
 } from "@app/lib/resources/skill/code_defined/shared";
+import { discoverKnowledgeSkill } from "@app/lib/resources/skill/code_defined/system/discover_knowledge";
+import { discoverSkillsSkill } from "@app/lib/resources/skill/code_defined/system/discover_skills";
+import { discoverToolsSkill } from "@app/lib/resources/skill/code_defined/system/discover_tools";
+import { planModeSkill } from "@app/lib/resources/skill/code_defined/system/plan_mode";
 import type { AllSkillConfigurationFindOptions } from "@app/lib/resources/skill/types";
 
 // Registry is a simple array.
@@ -16,7 +16,7 @@ const SYSTEM_SKILLS_ARRAY = ensureUniqueSIds([
   discoverSkillsSkill,
   discoverToolsSkill,
   planModeSkill,
-] as const);
+] as const satisfies readonly SystemSkillDefinition[]);
 
 // Build lookup map for direct access by sId.
 const SYSTEM_SKILLS_BY_ID: Map<string, SystemSkillDefinition> = new Map(
